Move find.rs into a directory
